+++ src/main/resources/src/components/MainPage/WorkshopLEMZ/TableView/TableView.jsx
@@ -149,7 +149,7 @@ const TableView = (props) => {
                     <div className="tableview_requests_LEMZ__actions">
                         <Link to={"/workshop-lemz/view/" + request.id} className="tableview_requests_LEMZ__action" >Просмотр</Link>
                         {props.userHasAccess(['ROLE_ADMIN', 'ROLE_MANAGER']) && <Link to={"/workshop-lemz/edit/" + request.id} className="tableview_requests_LEMZ__action">Редактировать</Link>}
-                        {/* <div data-id={request.id} className="tableview_requests_LEMZ__action" onClick={props.deleteItem}>Удалить</div> */}
+                        {props.userHasAccess(['ROLE_ADMIN']) && <div data-id={request.id} className="tableview_requests_LEMZ__action" onClick={props.deleteItem}>Удалить</div>}
                     </div>
                 </div>
             ))}
